personas/marketing/requirements.txt
streamlit==1.40.2
openai==1.55.1
Pillow==11.0.0
python-dotenv==1.0.1
watchdog==3.0.0
httpx==0.27.2  # Pinned version to fix OpenAI client compatibility